+++ python/mujoco/introspect/functions.py
@@ -10823,6 +10823,22 @@ FUNCTIONS: Mapping[str, FunctionDecl] = dict([
          ),
          doc='Get spec from body.',
      )),
+    ('mjs_getCompiler',
+     FunctionDecl(
+         name='mjs_getCompiler',
+         return_type=PointerType(
+             inner_type=ValueType(name='mjsCompiler'),
+         ),
+         parameters=(
+             FunctionParameterDecl(
+                 name='element',
+                 type=PointerType(
+                     inner_type=ValueType(name='mjsElement'),
+                 ),
+             ),
+         ),
+         doc="Get compiler associated with element's origin spec.",
+     )),
     ('mjs_findSpec',
      FunctionDecl(
          name='mjs_findSpec',